- Upgraded to scintilla 2.29 - Removed custom changes made to scintilla (escpecially the custom folding styles) - Updated the "Global Editore Preferences | Folding" page - Editor's virtual spaces is now accessbile only for rectangular selection
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5114 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ sdk/wxscintilla/src/scintilla/lexers/LexMagik.cxx
@@ -17,7 +17,6 @@
 #include "Scintilla.h"
 #include "SciLexer.h"
 
-#include "PropSetSimple.h"
 #include "WordList.h"
 #include "LexAccessor.h"
 #include "Accessor.h"
@@ -234,7 +233,7 @@ static void ColouriseMagikDoc(unsigned int startPos, int length, int initStyle,
 	            }
 
 	            if(characters.InList(keyword)) {
-	                sc.Forward(strlen(keyword));
+	                sc.Forward(static_cast<int>(strlen(keyword)));
 	            } else {
 	                sc.Forward();
 	            }
